+++ Modular Gameplay Overhaul/Modules/Core/Commands/RevalidateCommand.cs
@@ -2,8 +2,8 @@
 
 #region using directives
 
+using System.Collections.Generic;
 using System.Linq;
-using Arsenal;
 using DaLion.Overhaul;
 using DaLion.Overhaul.Modules.Arsenal.Extensions;
 using DaLion.Shared.Commands;
@@ -11,22 +11,24 @@ using DaLion.Shared.Extensions.Collections;
 using DaLion.Shared.Extensions.SMAPI;
 using DaLion.Shared.Extensions.Stardew;
 using StardewValley;
+using StardewValley.Locations;
+using StardewValley.Objects;
 using StardewValley.Tools;
 
 #endregion using directives
 
 [UsedImplicitly]
-internal sealed class RevalidateItemsCommand : ConsoleCommand
+internal sealed class RevalidateCommand : ConsoleCommand
 {
-    /// <summary>Initializes a new instance of the <see cref="RevalidateItemsCommand"/> class.</summary>
+    /// <summary>Initializes a new instance of the <see cref="RevalidateCommand"/> class.</summary>
     /// <param name="handler">The <see cref="CommandHandler"/> instance that handles this command.</param>
-    internal RevalidateItemsCommand(CommandHandler handler)
+    internal RevalidateCommand(CommandHandler handler)
         : base(handler)
     {
     }
 
     /// <inheritdoc />
-    public override string[] Triggers { get; } = { "revalidate_items", "revalidate", "reval", "rev" };
+    public override string[] Triggers { get; } = { "revalidate", "reval", "rev" };
 
     /// <inheritdoc />
     public override string Documentation => "Applies or removes persistent changes made by modules to existing items.";
@@ -55,8 +57,32 @@ internal sealed class RevalidateItemsCommand : ConsoleCommand
             Game1.player.WriteIfNotExists(DataFields.GalaxyArsenalObtained, Constants.GalaxySwordIndex.ToString());
         }
 
+        var removed = 0;
+        if (ArsenalModule.IsEnabled)
+        {
+            foreach (var chest in IterateAllChests())
+            {
+                while (chest.items.FirstOrDefault(
+                           i => i is MeleeWeapon { InitialParentTileIndex: Constants.DarkSwordIndex }) is { } darkSword)
+                {
+                    chest.items.Remove(darkSword);
+                    removed++;
+                }
+            }
+        }
+
         Log.I(
             $"All {(Context.IsMainPlayer ? "global" : "local")} items have been revalidated according to the current configuration settings.");
+        if (removed <= 0)
+        {
+            return;
+        }
+
+        Log.W($"{removed} Dark Swords were removed from Chests.");
+        if (Game1.player.hasOrWillReceiveMail("viegoCurse") && Game1.player.Items.FirstOrDefault(i => i is MeleeWeapon {InitialParentTileIndex: Constants.DarkSwordIndex}) is null && !Game1.player.addItemToInventoryBool(new MeleeWeapon(Constants.DarkSwordIndex)))
+        {
+            Log.E($"Failed adding Dark Sword to {Game1.player.Name}. Use CJB Item Spawner to obtain a new copy.");
+        }
     }
 
     private static void RevalidateSingleWeapon(MeleeWeapon weapon)
@@ -86,6 +112,46 @@ internal sealed class RevalidateItemsCommand : ConsoleCommand
             || weapon.InitialParentTileIndex is Constants.DarkSwordIndex or Constants.HolyBladeIndex))
         {
             weapon.specialItem = true;
+        }
+    }
+
+    private static IEnumerable<Chest> IterateAllChests()
+    {
+        foreach (var location1 in Game1.locations)
+        {
+            foreach (var @object in location1.objects.Values)
+            {
+                if (@object is Chest chest1)
+                {
+                    yield return chest1;
+                }
+                else if (@object.heldObject.Value is Chest chest2)
+                {
+                    yield return chest2;
+                }
+            }
+
+            if (location1 is not BuildableGameLocation buildable)
+            {
+                continue;
+            }
+
+            foreach (var location2 in buildable.buildings
+                         .Where(b => b.indoors.Value is not null)
+                         .Select(b => b.indoors.Value))
+            {
+                foreach (var @object in location2.objects.Values)
+                {
+                    if (@object is Chest chest1)
+                    {
+                        yield return chest1;
+                    }
+                    else if (@object.heldObject.Value is Chest chest2)
+                    {
+                        yield return chest2;
+                    }
+                }
+            }
         }
     }
 }