+++ Modular Gameplay Overhaul/Modules/Professions/Integrations/CustomOreNodesIntegration.cs
@@ -4,7 +4,6 @@
 
 using System.Linq;
 using DaLion.Shared.Attributes;
-using DaLion.Shared.Extensions.Reflection;
 using DaLion.Shared.Integrations;
 using DaLion.Shared.Integrations.CustomOreNodes;
 
@@ -18,32 +17,18 @@ internal sealed class CustomOreNodesIntegration : ModIntegration<CustomOreNodesI
     {
     }
 
-    /// <inheritdoc />
-    protected override bool RegisterImpl()
+    internal void RegisterCustomOreData()
     {
         if (!this.IsLoaded)
         {
-            return false;
+            return;
         }
 
-        try
-        {
-            var getCustomOreNodeParentSheetIndex =
-                "CustomOreNodes.CustomOreNode"
-                    .ToType()
-                    .RequireField("parentSheetIndex");
-            Collections.ResourceNodeIds = Collections.ResourceNodeIds
-                .Concat(
-                    this.ModApi
-                        .GetCustomOreNodes()
-                        .Select(n => (int)getCustomOreNodeParentSheetIndex.GetValue(n)!))
-                .ToHashSet();
-            return true;
-        }
-        catch (Exception ex)
-        {
-            Log.W($"Could not retrieve custom ore node IDs.\n{ex}");
-            return false;
-        }
+        Collections.ResourceNodeIds = Collections.ResourceNodeIds
+            .Concat(
+                this.ModApi
+                    .GetCustomOreNodes()
+                    .Select(n => Reflector.GetUnboundFieldGetter<object, int>(n, "parentSheetIndex").Invoke(n)))
+            .ToHashSet();
     }
 }